split UI agnostic code from font.lua into fontlist.lua
